+++ skrub/_utils.py
@@ -4,8 +4,10 @@ import secrets
 from typing import Iterable
 
 import numpy as np
+import sklearn
 from sklearn.base import clone
 from sklearn.utils import check_array
+from sklearn.utils.fixes import parse_version
 
 from skrub import _dataframe as sbd
 
@@ -38,6 +40,37 @@ class LRUDict:
 
     def __contains__(self, key):
         return key in self.cache
+
+
+def unique_strings(values, is_null):
+    """Unique values, accounting for nulls.
+
+    This is like np.unique except
+    - it is only for 1d arrays of strings
+    - caller must pass a boolean array indicating which values are null: ``is_null``
+    - null values are considered to be the same as the empty string.
+
+    >>> import numpy as np
+    >>> import pandas as pd
+    >>> from skrub._utils import unique_strings
+    >>> a = np.asarray(['paris', '', 'berlin', None, 'london'])
+    >>> values, idx = unique_strings(a, pd.isna(a))
+    >>> values
+    array(['', 'berlin', 'london', 'paris'], dtype=object)
+    >>> values[idx]
+    array(['paris', '', 'berlin', '', 'london'], dtype=object)
+    """
+    not_null_values = values[~is_null]
+    unique, idx = np.unique(not_null_values, return_inverse=True)
+    if not is_null.any():
+        return unique, idx
+    if not len(unique) or unique[0] != "":
+        unique = np.concatenate([[""], unique])
+        idx += 1
+    full_idx = np.empty(values.shape, dtype=idx.dtype)
+    full_idx[is_null] = 0
+    full_idx[~is_null] = idx
+    return unique, full_idx
 
 
 def check_input(X):
@@ -154,22 +187,80 @@ def repr_args(args, kwargs, defaults={}):
     )
 
 
-def transformer_output_type_error(transformer, transform_input, transform_output):
-    module = sbd.dataframe_module_name(transform_input)
+def set_output(transformer, X):
+    if not hasattr(transformer, "set_output"):
+        return
+    module_name = sbd.dataframe_module_name(X)
+    if module_name == "polars" and parse_version(sklearn.__version__) < parse_version(
+        "1.4"
+    ):
+        # TODO: remove when scikit-learn 1.3 support is dropped.
+        transformer.set_output(transform="pandas")
+    else:
+        transformer.set_output(transform=module_name)
+
+
+def check_output(
+    transformer, transform_input, transform_output, allow_column_list=True
+):
+    target_module = sbd.dataframe_module_name(transform_input)
+
+    def has_correct_module(obj):
+        return sbd.dataframe_module_name(obj) == target_module
+
+    if (
+        sbd.is_dataframe(transform_output)
+        and target_module == "polars"
+        and sbd.dataframe_module_name(transform_output) == "pandas"
+        and hasattr(transformer, "set_output")
+        and parse_version(sklearn.__version__) < parse_version("1.4")
+    ):
+        # TODO: remove when scikit-learn 1.3 support is dropped.
+        #
+        # For older scikit-learn versions that do not support
+        # `set_output(transform='polars')`, we fall back to using
+        # `set_output(transform='pandas')` and converting the output dataframe
+        # to polars ourselves.
+        # Therefore having pandas output when the input is polars is tolerated,
+        # when:
+        #   - the scikit-learn version is < 1.4
+        #   - and the transformer relies on the set_output API
+        #     (this implies that the output is a dataframe -- not a column or
+        #     list of columns).
+        # In all other cases having the output backed by the wrong dataframe
+        # library (e.g. pandas instead of polars) will result in an error.
+
+        # transform_input is a polars object so we know we can import it
+        import polars as pl
+
+        return pl.from_pandas(transform_output)
+    if sbd.is_dataframe(transform_output) and has_correct_module(transform_output):
+        return transform_output
+    if (
+        allow_column_list
+        and sbd.is_column(transform_output)
+        and has_correct_module(transform_output)
+    ):
+        return transform_output
+    if (
+        allow_column_list
+        and sbd.is_column_list(transform_output)
+        and (not len(transform_output) or has_correct_module(transform_output[0]))
+    ):
+        return transform_output
     message = (
         f"{transformer.__class__.__name__}.fit_transform returned a result of type"
-        f" {transform_output.__class__.__name__}, but a {module} DataFrame was"
+        f" {transform_output.__class__.__name__}, but a {target_module} DataFrame was"
         f" expected. If {transformer.__class__.__name__} is a custom transformer class,"
-        f" please make sure that the output is a {module} container when the input is a"
-        f" {module} container."
+        f" please make sure that the output is a {target_module} container when the"
+        f" input is a {target_module} container."
     )
     if not hasattr(transformer, "set_output"):
         message += (
-            f" One way of enabling a transformer to output {module} DataFrames is"
-            " inheriting from the sklearn.base.TransformerMixin class and defining the"
-            " 'get_feature_names_out' method. See"
-            " https://scikit-learn.org/stable/auto_examples/"
-            "miscellaneous/plot_set_output.html"
+            f" One way of enabling a transformer to output {target_module} DataFrames"
+            " is inheriting from the sklearn.base.TransformerMixin class and defining"
+            " the 'get_feature_names_out' method. See"
+            " https://scikit-learn.org/stable/auto_examples/miscellaneous/plot_set_output.html"
             " for details."
         )
     raise TypeError(message)